(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -7,6 +7,10 @@ const controller = {}     // Objeto vazio
 controller.create = async function(req, res) {
   try {
 
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     // Verifica se existe o campo "password" em "req.body".
     // Caso positivo, geramos o hash da senha antes de enviá-la
     // ao BD
@@ -29,9 +33,13 @@ controller.create = async function(req, res) {
   }
 }
 
-
 controller.retrieveAll = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     const result = await prisma.user.findMany(
       // Omite o campo "password" do resultado
       // por questão de segurança
@@ -51,13 +59,20 @@ controller.retrieveAll = async function(req, res) {
 
 controller.retrieveOne = async function(req, res) {
   try {
+
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if(! (req?.authUser?.is_admin || 
+      Number(req?.authUser?.id) === Number(req.params.id))) 
+      return res.status(403).end()
+
     const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questão de segurança
       omit: { password: true },
       where: { id: Number(req.params.id) }
     })
-
 
     // Encontrou ~> retorna HTTP 200: OK (implícito)
     if(result) res.send(result)
@@ -74,6 +89,10 @@ controller.retrieveOne = async function(req, res) {
 
 controller.update = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
 
     // Verifica se existe o campo "password" em "req.body".
     // Caso positivo, geramos o hash da senha antes de enviá-la
@@ -102,9 +121,13 @@ controller.update = async function(req, res) {
   }
 }
 
-
 controller.delete = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
@@ -126,7 +149,6 @@ controller.delete = async function(req, res) {
     }
   }
 }
-
 
 controller.login = async function(req, res) {
   try {
@@ -158,6 +180,10 @@ controller.login = async function(req, res) {
       // Se a senha estiver errada, retorna
       // HTTP 401: Unauthorized
       if(! passwordIsValid) return res.status(401).end()
+
+      // Por motivos de segurança, exclui o campo "password" dos dados do usuário
+      // para que ele não seja incluído no token
+      if(user.password) delete user.password
 
       // Usuário e senha OK, passamos ao procedimento de gerar o token
       const token = jwt.sign(
